Refactor batch_iterable() function into utility file
