+++ src/biochemical_data_connectors/connectors/bioactive_compounds_connectors.py
@@ -10,12 +10,12 @@ import pubchempy as pcp
 from chembl_webresource_client.new_client import new_client
 
 from src.biochemical_data_connectors.constants import RestApiEndpoints
+from src.biochemical_data_connectors.utils.iter_utils import batch_iterable
 from src.biochemical_data_connectors.utils.api.mappings import uniprot_to_gene_id_mapping
 from src.biochemical_data_connectors.utils.api.pubchem_api import (
     get_active_aids,
     get_active_cids,
     get_compounds_in_batches,
-    batch_iterable,
     get_compound_potency
 )
